fix(repository): Fix get limit return type

+++ app/Repositories/MachineLearningFacade.php
@@ -6,5 +6,5 @@ interface MachineLearningFacade
 {
     public function getPersonality($features): string;
 
-    public function getLimit($features): int;
+    public function getLimit($features): float;
 }
